feat: uploaded FordGo bike analysis notebooks
Ford Go Bike Part_I_exploration is a jupyter notebook where I explored a dataset of  a bike sharing system in the United States 
using python data science and visualization libraries.  
The goal of the exploration is to provide a revamped dataset for 
effective analysis and to create create  key insights for users.